refactor(webapp): extract DOM-free account-store from ui/provider-settings (#968)
Split the data layer out of the 2,577-line DOM-heavy
ui/provider-settings.ts into a new DOM-free providers/account-store.ts.
Worker-resident scoops, shell commands, and built-in providers now import
the pure account/model accessors from account-store instead of reaching up
into the ui/ layer, so the settings dialog's DOM code is tree-shake-eligible
from the kernel-worker bundle. ui/provider-settings.ts keeps the dialog and
re-exports the data layer for backward compatibility.
