+++ packages/webapp/tests/scoops/scoop-context.restore-orphans.test.ts
@@ -66,7 +66,7 @@ vi.mock('../../src/shell/index.js', () => ({
   }),
 }));
 
-vi.mock('../../src/ui/provider-settings.js', () => ({
+vi.mock('../../src/providers/account-store.js', () => ({
   getApiKey: () => 'test-api-key',
   getSelectedProvider: () => 'anthropic',
   resolveCurrentModel: () => ({ id: 'test-model', provider: 'anthropic' }),
