+++ packages/webapp/tests/scoops/scoop-context.session-id.test.ts
@@ -89,7 +89,7 @@ vi.mock('../../src/shell/index.js', () => ({
   }),
 }));
 
-vi.mock('../../src/ui/provider-settings.js', () => ({
+vi.mock('../../src/providers/account-store.js', () => ({
   getApiKey: () => 'test-api-key',
   getSelectedProvider: () => 'anthropic',
   resolveCurrentModel: mocks.resolveCurrentModel,
