+++ packages/webapp/tests/shell/supplemental-commands/models-command.current-model.test.ts
@@ -8,7 +8,7 @@
 import type { IFileSystem } from 'just-bash';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 
-vi.mock('../../../src/ui/provider-settings.js', () => ({
+vi.mock('../../../src/providers/account-store.js', () => ({
   getAccounts: vi.fn(),
   getAvailableProviders: vi.fn(),
   getProviderConfig: vi.fn(),
@@ -18,7 +18,6 @@ vi.mock('../../../src/ui/provider-settings.js', () => ({
   resolveCurrentModel: vi.fn(),
 }));
 
-import { createModelsCommand } from '../../../src/shell/supplemental-commands/models-command.js';
 import {
   getAccounts,
   getAvailableProviders,
@@ -27,7 +26,8 @@ import {
   getSelectedModelId,
   getSelectedProvider,
   resolveCurrentModel,
-} from '../../../src/ui/provider-settings.js';
+} from '../../../src/providers/account-store.js';
+import { createModelsCommand } from '../../../src/shell/supplemental-commands/models-command.js';
 
 const mk = (id: string, costIn: number, costOut: number, ctx = 1_000_000, provider = 'adobe') => ({
   id,
